fixed all warnings

+++ Utopia/libs/librabbitmq/amqp_connection.c
@@ -245,7 +245,7 @@ int amqp_handle_input(amqp_connection_state_t state,
   /* do we have target_size data yet? if not, return with the
      expectation that more will arrive */
   if (state->inbound_offset < state->target_size) {
-    return bytes_consumed;
+    return (int)bytes_consumed;
   }
 
   raw_frame = state->inbound_buffer.bytes;
@@ -267,7 +267,7 @@ int amqp_handle_input(amqp_connection_state_t state,
         = amqp_d8(raw_frame, 7);
 
       return_to_idle(state);
-      return bytes_consumed;
+      return (int)bytes_consumed;
     }
 
     /* it's not a protocol header; fall through to process it as a
@@ -305,7 +305,7 @@ int amqp_handle_input(amqp_connection_state_t state,
     /* do we have target_size data yet? if not, return with the
        expectation that more will arrive */
     if (state->inbound_offset < state->target_size) {
-      return bytes_consumed;
+      return (int)bytes_consumed;
     }
 
   }
@@ -380,7 +380,7 @@ int amqp_handle_input(amqp_connection_state_t state,
     }
 
     return_to_idle(state);
-    return bytes_consumed;
+    return (int)bytes_consumed;
   }
 
   default:
@@ -457,7 +457,7 @@ int amqp_send_frame(amqp_connection_state_t state,
     uint8_t frame_end_byte = AMQP_FRAME_END;
     const amqp_bytes_t *body = &frame->payload.body_fragment;
 
-    amqp_e32(out_frame, 3, body->len);
+    amqp_e32(out_frame, 3, (int)body->len);
 
     iov[0].iov_base = out_frame;
     iov[0].iov_len = HEADER_SIZE;
@@ -466,7 +466,7 @@ int amqp_send_frame(amqp_connection_state_t state,
     iov[2].iov_base = &frame_end_byte;
     iov[2].iov_len = FOOTER_SIZE;
 
-    res = amqp_socket_writev(state->socket, iov, 3);
+    res = (int)amqp_socket_writev(state->socket, iov, 3);
   } else {
     size_t out_frame_len;
     amqp_bytes_t encoded;
@@ -512,9 +512,9 @@ int amqp_send_frame(amqp_connection_state_t state,
       return AMQP_STATUS_INVALID_PARAMETER;
     }
 
-    amqp_e32(out_frame, 3, out_frame_len);
+    amqp_e32(out_frame, 3, (int)out_frame_len);
     amqp_e8(out_frame, out_frame_len + HEADER_SIZE, AMQP_FRAME_END);
-    res = amqp_socket_send(state->socket, out_frame,
+    res = (int)amqp_socket_send(state->socket, out_frame,
                            out_frame_len + HEADER_SIZE + FOOTER_SIZE);
   }
